Update g3read.py
The test on overflow is wrong. It assumes that the blocks are written to all particle types, which is false. One can only check if we have an overflow problem if the info block is written and we know to which particle types the block has been written. But we have a vicious circle here as the function is called when the info is not read. I could not find a working solution for keeping the test while correcting the bug.